# aspnetcore/src/IO.Swagger/Controllers/ApplicationInformationApi.cs
/*
 * Green Button Resource Server API Documentation
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * Contact: dcoffin@greenbuttonalliance.org
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class ApplicationInformationApiController : ControllerBase
    { 
        /// <summary>
        /// List of all ApplicationInformation resources
        /// </summary>
        /// <response code="200">OK</response>
        /// <response code="400">Bad Request</response>
        /// <response code="401">Unauthorized</response>
        /// <response code="403">Forbidden</response>
        /// <response code="404">Not Found</response>
        [HttpGet]
        [Route("/GreenButtonAlliance/Green_Button_Resource_Server/espi/1_1/resource/ApplicationInformation")]
        [ValidateModelState]
        [SwaggerOperation("GetApplicationInformation")]
        [SwaggerResponse(statusCode: 200, type: typeof(ApplicationInformation), description: "OK")]
        public virtual IActionResult GetApplicationInformation()
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(ApplicationInformation));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            //TODO: Uncomment the next line to return response 401 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(401);

            //TODO: Uncomment the next line to return response 403 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(403);

            //TODO: Uncomment the next line to return response 404 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(404);
            string exampleJson = null;
            exampleJson = "{\n  \"thirdPartyApplicationType\" : 1,\n  \"thirdPartyNotifyUri\" : \"https://sandbox.greenbuttonalliance.org:8443/ThirdParty/espi/1_1/Notification\",\n  \"client_uri\" : \"https://sandbox.greenbuttonalliance.org:8443/ThirdParty\",\n  \"dataCustodianBulkRequestURI\" : \"https://sandbox.greenbuttonalliance.org:8443/DataCustodian/espi/1_1/resource/Batch/Bulk/{bulkId}\",\n  \"authorizationServerUri\" : \"https://sandbox.greenbuttonalliance.org:8443/DataCustodian\",\n  \"dataCustodianApplicationStatus\" : 2,\n  \"authorizationServerRegistrationEndpoint\" : \"https://sandbox.greenbuttonalliance.org:8443/DataCustodian/oauth/register\",\n  \"logo_uri\" : \"https://sandbox.greenbuttonalliance.org:8443/ThirdParty/resources/ico/favicon.png\",\n  \"authorizationServerAuthorizationEndpoint\" : \"https://sandbox.greenbuttonalliance.org:8443/DataCustodian/oauth/authorize\",\n  \"thirdPartyPhone\" : \"(123) 456-7890\",\n  \"client_id\" : \"641d16dcbacb0e40e41d\",\n  \"dataCustodianId\" : \"817212172021\",\n  \"dataCustodianResourceEndpoint\" : \"https://sandbox.greenbuttonalliance.org:8443/DataCustodian/espi/1_1/resource\",\n  \"thirdPartyApplicationStatus\" : 3,\n  \"thirdPartyApplicationDescription\" : \"The SolarPower application helps electricity customers to reduce their energy expenses by converting to solar energy.\",\n  \"thirdPartyApplicationUse\" : 1,\n  \"authorizationServerTokenEndpoint\" : \"https://sandbox.greenbuttonalliance.org:8443/DataCustodian/oauth/token\",\n  \"client_secret\" : \"secret\",\n  \"redirect_uri\" : \"https://sandbox.greenbuttonalliance.org:8443/ThirdParty/espi/1_1/OAuthCallBack\",\n  \"thirdPartyUserPortalScreenURI\" : \"https://sandbox.greenbuttonalliance.org:8443/ThirdParty/\",\n  \"client_name\" : \"Green Button Alliance, Inc. Demonstration Third Party Application\",\n  \"thirdPartyScopeSelectionURI\" : \"https://sandbox.greenbuttonalliance.org:8443/ThirdParty/RetailCustomer/ScopeSelection\"\n}";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<ApplicationInformation>(exampleJson)
                        : default(ApplicationInformation);            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}