# aspnetcore/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "Green Button Resource Server API Documentation",
    "contact" : {
      "name" : "Donald F. Coffin",
      "url" : "https://greenbuttonalliance.org",
      "email" : "dcoffin@greenbuttonalliance.org"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/GreenButtonAlliance/Green_Button_Resource_Server",
    "description" : "SwaggerHub API Auto Mocking"
  }, {
    "url" : "https://sandbox.greenbuttonalliance.org:8443/DataCustodian",
    "description" : "Green Button Alliance (GBA) Production ESPI Resource Server Sandbox"
  }, {
    "url" : "https://sandboxdev.greenbuttonalliance.org:8443/DataCustodian",
    "description" : "Green Button Alliance (GBA) Development ESPI Resource Server Sandbox"
  } ],
  "paths" : {
    "/espi/1_1/resource/ApplicationInformation" : {
      "get" : {
        "tags" : [ "ApplicationInformation" ],
        "summary" : "List of all ApplicationInformation resources",
        "operationId" : "getApplicationInformation",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/atom+xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApplicationInformation"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "ApplicationInformation" : {
        "required" : [ "authorizationServerAuthorizationEndpoint", "authorizationServerTokenEndpoint", "client_id", "client_name", "client_secret", "dataCustodianApplicationStatus", "dataCustodianBulkRequestURI", "dataCustodianId", "dataCustodianResourceEndpoint", "redirect_uri", "thirdPartyNotifyUri" ],
        "type" : "object",
        "properties" : {
          "dataCustodianId" : {
            "maxLength" : 64,
            "type" : "string",
            "description" : "Data Custodian Id",
            "example" : "817212172021"
          },
          "dataCustodianApplicationStatus" : {
            "type" : "integer",
            "description" : "Code indicating the current status of the Data Custodian application",
            "example" : 2,
            "enum" : [ 1, 2, 3, 4 ]
          },
          "thirdPartyApplicationDescription" : {
            "maxLength" : 256,
            "type" : "string",
            "description" : "Description of the Third Party application",
            "example" : "The SolarPower application helps electricity customers to reduce their energy expenses by converting to solar energy."
          },
          "thirdPartyApplicationStatus" : {
            "type" : "integer",
            "description" : "Code indicating the current status of the Third Party application",
            "example" : 3,
            "enum" : [ 1, 2, 3, 4 ]
          },
          "thirdPartyApplicationType" : {
            "type" : "integer",
            "description" : "Code indicating the Third Party application type",
            "example" : 1,
            "enum" : [ 1, 2, 3, 4 ]
          },
          "thirdPartyApplicationUse" : {
            "type" : "integer",
            "description" : "Code indicating the expected use of the Third Party application",
            "example" : 1,
            "enum" : [ 1, 2, 3, 4, 5 ]
          },
          "thirdPartyPhone" : {
            "maxLength" : 32,
            "type" : "string",
            "description" : "Third Party Application Technical Support phone number",
            "example" : "(123) 456-7890"
          },
          "authorizationServerUri" : {
            "type" : "string",
            "description" : "Base URI of the OAuth 2.0 Authorization Server",
            "format" : "uri",
            "example" : "https://sandbox.greenbuttonalliance.org:8443/DataCustodian"
          },
          "thirdPartyNotifyUri" : {
            "type" : "string",
            "description" : "URI value used to notify the Third Party application the Data Custodian has data available",
            "format" : "uri",
            "example" : "https://sandbox.greenbuttonalliance.org:8443/ThirdParty/espi/1_1/Notification"
          },
          "authorizationServerAuthorizationEndpoint" : {
            "type" : "string",
            "description" : "OAuth 2.0 Authorization Server's Authorization Endpoint",
            "format" : "uri",
            "example" : "https://sandbox.greenbuttonalliance.org:8443/DataCustodian/oauth/authorize"
          },
          "authorizationServerRegistrationEndpoint" : {
            "type" : "string",
            "description" : "OAuth 2.0 Authorization Server's Dynamic Client Registration Endpoint",
            "format" : "uri",
            "example" : "https://sandbox.greenbuttonalliance.org:8443/DataCustodian/oauth/register"
          },
          "authorizationServerTokenEndpoint" : {
            "type" : "string",
            "description" : "OAuth 2.0 Authorization Server's Token Endpoint",
            "format" : "uri",
            "example" : "https://sandbox.greenbuttonalliance.org:8443/DataCustodian/oauth/token"
          },
          "dataCustodianBulkRequestURI" : {
            "type" : "string",
            "description" : "URI used by the Data Custodian to support Bulk API Request",
            "format" : "uri",
            "example" : "https://sandbox.greenbuttonalliance.org:8443/DataCustodian/espi/1_1/resource/Batch/Bulk/{bulkId}"
          },
          "dataCustodianResourceEndpoint" : {
            "type" : "string",
            "description" : "Base value of Data Custodian Resource API URIs",
            "format" : "uri",
            "example" : "https://sandbox.greenbuttonalliance.org:8443/DataCustodian/espi/1_1/resource"
          },
          "thirdPartyScopeSelectionURI" : {
            "type" : "string",
            "description" : "\\[DEPRECATED\\] URI of the Third Party's Scope Selection API used by Data Custodian Applications to initiate OAuth 2.0 Authorization Code Request",
            "format" : "uri",
            "example" : "https://sandbox.greenbuttonalliance.org:8443/ThirdParty/RetailCustomer/ScopeSelection"
          },
          "thirdPartyUserPortalScreenURI" : {
            "type" : "string",
            "description" : "URI of the Third Party Application's starting screen",
            "format" : "uri",
            "example" : "https://sandbox.greenbuttonalliance.org:8443/ThirdParty/"
          },
          "client_secret" : {
            "maxLength" : 512,
            "type" : "string",
            "description" : "OAuth 2.0 assigned **secret** provided by the Data Custodian during the Third Party Application registration process.",
            "example" : "secret"
          },
          "logo_uri" : {
            "type" : "string",
            "description" : "URI of Third Party logo",
            "format" : "uri",
            "example" : "https://sandbox.greenbuttonalliance.org:8443/ThirdParty/resources/ico/favicon.png"
          },
          "client_name" : {
            "maxLength" : 256,
            "type" : "string",
            "description" : "The name of the registered Third Party Application",
            "example" : "Green Button Alliance, Inc. Demonstration Third Party Application"
          },
          "client_uri" : {
            "type" : "string",
            "description" : "URI of the Third Party Application's starting screen",
            "format" : "uri",
            "example" : "https://sandbox.greenbuttonalliance.org:8443/ThirdParty"
          },
          "redirect_uri" : {
            "type" : "array",
            "description" : "Third Party Application's default OAuth 2.0 **Authorization Code** Request redirect_uri parameter value",
            "example" : "https://sandbox.greenbuttonalliance.org:8443/ThirdParty/espi/1_1/OAuthCallBack",
            "items" : {
              "type" : "string",
              "format" : "uri"
            }
          },
          "client_id" : {
            "maxLength" : 64,
            "type" : "string",
            "description" : "Third Party Application Unique Identifier provided by the Data Custodian during the Third Party Application registration process.",
            "example" : "641d16dcbacb0e40e41d"
          }
        },
        "example" : {
          "thirdPartyApplicationType" : 1,
          "thirdPartyNotifyUri" : "https://sandbox.greenbuttonalliance.org:8443/ThirdParty/espi/1_1/Notification",
          "client_uri" : "https://sandbox.greenbuttonalliance.org:8443/ThirdParty",
          "dataCustodianBulkRequestURI" : "https://sandbox.greenbuttonalliance.org:8443/DataCustodian/espi/1_1/resource/Batch/Bulk/{bulkId}",
          "authorizationServerUri" : "https://sandbox.greenbuttonalliance.org:8443/DataCustodian",
          "dataCustodianApplicationStatus" : 2,
          "authorizationServerRegistrationEndpoint" : "https://sandbox.greenbuttonalliance.org:8443/DataCustodian/oauth/register",
          "logo_uri" : "https://sandbox.greenbuttonalliance.org:8443/ThirdParty/resources/ico/favicon.png",
          "authorizationServerAuthorizationEndpoint" : "https://sandbox.greenbuttonalliance.org:8443/DataCustodian/oauth/authorize",
          "thirdPartyPhone" : "(123) 456-7890",
          "client_id" : "641d16dcbacb0e40e41d",
          "dataCustodianId" : "817212172021",
          "dataCustodianResourceEndpoint" : "https://sandbox.greenbuttonalliance.org:8443/DataCustodian/espi/1_1/resource",
          "thirdPartyApplicationStatus" : 3,
          "thirdPartyApplicationDescription" : "The SolarPower application helps electricity customers to reduce their energy expenses by converting to solar energy.",
          "thirdPartyApplicationUse" : 1,
          "authorizationServerTokenEndpoint" : "https://sandbox.greenbuttonalliance.org:8443/DataCustodian/oauth/token",
          "client_secret" : "secret",
          "redirect_uri" : "https://sandbox.greenbuttonalliance.org:8443/ThirdParty/espi/1_1/OAuthCallBack",
          "thirdPartyUserPortalScreenURI" : "https://sandbox.greenbuttonalliance.org:8443/ThirdParty/",
          "client_name" : "Green Button Alliance, Inc. Demonstration Third Party Application",
          "thirdPartyScopeSelectionURI" : "https://sandbox.greenbuttonalliance.org:8443/ThirdParty/RetailCustomer/ScopeSelection"
        }
      }
    }
  }
}